feat: three different tools for summarization created (for files, file paths and url)

+++ libs/text_utils.py
@@ -5,21 +5,28 @@ from langchain.text_splitter import CharacterTextSplitter
 from docx import Document
 
 
-def get_text(doc):
-    # Create a temporary file
-    with tempfile.NamedTemporaryFile(delete=False) as temp:
-        # Write the uploaded file data to the temporary file
-        temp.write(doc.read())
-        # Get the path of the temporary file
-        temp_path = temp.name
+def get_text_from_file(file):
+    text = ""
 
-    # Determine the file extension
-    extension = doc.name.split('.')[-1]
-    # Extract the text from the temporary file using textract
-    text = textract.process(temp_path, extension=extension).decode('utf-8')
+    # If the file is a PDF, use the PyPDF2 library to extract the text
+    if file.type == "application/pdf":
+        text = get_pdf_text(file)
+    else:
+        # Other formats allowed by textract: doc, docx, eml, epub, html, json, rtf, txt, odt
+        # TODO Test all the textract formats
 
-    # Delete the temporary file
-    os.remove(temp_path)
+        # Create a temporary file
+        with tempfile.NamedTemporaryFile(delete=False) as temp:
+            # Write the uploaded file data to the temporary file
+            temp.write(file.read())
+            # Get the path of the temporary file
+            temp_path = temp.name
+
+        # Extract the text from the temporary file
+        text = extract_text(temp_path)
+
+        # Delete the temporary file
+        os.remove(temp_path)
 
     return text
 
@@ -83,3 +90,24 @@ def get_text_segments(source: str, sourceType: str) -> list[str]:
 
     return segments
 
+
+def get_text_from_path(path: str) -> str:
+    text = ""
+    if path.endswith(".pdf"):
+        text = get_pdf_text(path)
+    else:
+        text = extract_text(path)
+
+    return text
+
+
+def extract_text(file_path: str) -> str:
+    """
+    Extracts text from a file at the given path.
+    """
+
+    # Determine the file extension
+    extension = file_path.split('.')[-1]
+    # Extract the text from the temporary file using textract
+    text = textract.process(file_path, extension=extension).decode('utf-8')
+    return text
